Cleaned up appsettings, moved CompilerHostTokenService to RuntimeHost project
