+++ src/Diagnostics.CompilerHost/appsettings.json
@@ -13,8 +13,8 @@
     }
   },
   "AzureAd": {
-    "AADAuthority": "https://login.microsoftonline.com/microsoft.onmicrosoft.com",
-    "ClientId": "cda572d8-1f3e-4119-9a11-0e4ceb2fb3e1", // Azure AAd App id of CompilerHost
-    "AllowedAppIds": "b9a1efcd-32ee-4330-834c-c04eb00f4b33" // Azure AAD App id of RuntimeHost
+    "AADAuthority": "",
+    "ClientId": "", // Azure AAd App id of CompilerHost
+    "AllowedAppIds": "" // Azure AAD App id of RuntimeHost
   }
 }
