+++ src/Diagnostics.RuntimeHost/appsettings.json
@@ -13,7 +13,11 @@
     }
   },
   "CompilerHost": {
-    "CompilerHostUrl": "http://localhost:7000"
+    "CompilerHostUrl": "http://localhost:7000",
+    "ClientId": "",
+    "AppKey": "",
+    "AADAuthority": "",
+    "AADResource": ""
   },
   "SourceWatcher": {
     "WatcherType": "",
@@ -82,11 +86,5 @@
   },
   "SearchAPI": {
     "SearchAPIEnabled": false
-  },
-  "CompilerHostAuth": {
-    "ClientId": "b9a1efcd-32ee-4330-834c-c04eb00f4b33", // Azure AAD App id of RuntimeHost
-    "AppKey": "", // Client secret used to sign the token
-    "AADAuthority": "https://login.microsoftonline.com/microsoft.onmicrosoft.com",
-    "AADResource": "cda572d8-1f3e-4119-9a11-0e4ceb2fb3e1"// Azure AAD App id of CompilerHost
   }
 }
